llm: New memory management
This changes the memory allocation strategy from upfront estimation to
tracking actual allocations done by the engine and reacting to that. The
goal is avoid issues caused by both under-estimation (crashing) and
over-estimation (low performance due to under-utilized GPUs).

It is currently opt-in and can be enabled for models running on the
Ollama engine by setting OLLAMA_NEW_ESTIMATES=1. Behavior in other
cases is unchanged and will continue to use the existing estimates.

Further optimization is needed before making this the default.
Current limitations are:
 * Slower startup time in some cases
 * Potential transient overloading of 1 GPU before stabilizing (especially on Windows)
 * Context shifting or defrag may create a compute graph larger than the forward pass
   graph, leading to a crash
 * Models with large vision graphs may not allocate the maximum number of layers
   possible if the entire model does not fit